cfg80211: change NAN security context length to u16
NAN security context length field is defined in NAN spec (r34) as
2 octets long, so u16 should be used for the security context length.

type=feature
ticket=none

Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/179368
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: 283059c8093e338c94599cffd105d5b0d639d102

+++ include/net/cfg80211.h
@@ -2893,7 +2893,7 @@ struct cfg80211_nan_func_filter {
  */
 struct cfg80211_nan_sec_ctx_id {
 	enum nl80211_nan_sec_ctx_type type;
-	u8 len;
+	u16 len;
 	const u8 *data;
 };
